requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
django-mathfilters==1.0.0
razorpay==1.4.1
pytesseract==0.3.10
waitress==3.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
